More work on cascade mode
# Changes:

-  adjusted a bunch of divisions to multiplications by pre-computed inverses
-  also cached a bunch of 2*mass to be stored in the atom object itself
-  re-arranged where site force resetting is done, and which flag trigger it
-  added some more columns to the data file outputs, related to energy error
-  adjusted how the criteria for spring breaking was checked

# New:

-  Added energy conservation criteria for the integrator
-  Adjusted the failed DE parameter to control the maximum allowed error per step
-  Kinetic and Potential energy trackers are added to the site objects, as well as the lattice object, these trackers are reset at the beginning of each integration step, and as such, only consider the energy in the currently tracked set of particles

+++ src/safio.h
@@ -84,7 +84,7 @@ struct Safio
     // Radial distance from the ion to search for cells of lattice sites
     int DIST_SEARCH;
     // Average change in energy over last 3 steps to consider a failure
-    int FAILED_DE;
+    double FAILED_DE;
 
     // Flag controlling whether scat or test mode, scat mode is 666
     int SCAT_FLAG;
